refactor: move subrace into own file
