+++ src/parsing/models/character_subrace_file_parser.hpp
@@ -7,6 +7,7 @@
 
 #include "controllers/groups.hpp"
 #include "models/character_race.hpp"
+#include "models/character_subrace.hpp"
 #include "parsing/models/feature_holder_file_parser.hpp"
 
 namespace dnd {
