+++ tests/parsing/models/character_file_parser_test.cpp
@@ -17,6 +17,7 @@
 #include "models/character_class.hpp"
 #include "models/character_race.hpp"
 #include "models/character_subclass.hpp"
+#include "models/character_subrace.hpp"
 #include "models/effect_holder/feature.hpp"
 #include "models/spell.hpp"
